Remove isNewSubscription param from initUniqueKeyCountHll
The method now always either restores from checkpoint or creates a
fresh HLL. The pre-deployment "do nothing" case is handled by the
caller not invoking the method at all. This moves the subscription
check to the SIT call site where the offset context is available.

+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/stats/IngestionStatsReporter.java
@@ -81,7 +81,7 @@ public class IngestionStatsReporter extends AbstractVeniceStatsReporter<Ingestio
             WRITE_COMPUTE_OPERATION_FAILURE));
 
     registerSensor(new IngestionStatsGauge(this, () -> getStats().getStorageQuotaUsed(), 0, STORAGE_QUOTA_USED));
-    registerSensor(new IngestionStatsGauge(this, () -> (double) getStats().getUniqueKeyCount(), 0, UNIQUE_KEY_COUNT));
+    registerSensor(new IngestionStatsGauge(this, () -> getStats().getUniqueKeyCount(), 0, UNIQUE_KEY_COUNT));
 
     registerSensor(
         new IngestionStatsGauge(this, () -> getStats().getRecordsConsumed(), 0, RECORDS_CONSUMED_METRIC_NAME));
